refactor: move settings/topics to admin/taxonomy
